% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_helpers.R
\name{add_edge}
\alias{add_edge}
\title{Adds an edge to a markov_chain}
\usage{
add_edge(m, from, to, prob = 1, ...)
}
\arguments{
\item{m}{a markov_chain object}

\item{from}{a state. If string then interpreted as a label, otherwise index.}

\item{to}{a state. If string then interpreted as a label, otherwise index.}

\item{prob}{the probability (defaults fo 1)}

\item{...}{Extra arguments like color, loop_angle, curve, etc.}
}
\value{
markov_chain object
}
\description{
Adds an edge to a markov_chain
}
