% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_class.R
\name{add_edge_by_indices}
\alias{add_edge_by_indices}
\title{Add an edge to the chain}
\usage{
add_edge_by_indices(
  m,
  from,
  to,
  prob = 1,
  label = NA,
  color = NA,
  curve = 0,
  loop_angle = 0
)
}
\arguments{
\item{m}{markov_chain}

\item{from}{string label}

\item{to}{string label}

\item{prob}{numeric}

\item{label}{string}

\item{color}{string}

\item{curve}{numeric}

\item{loop_angle}{numeric}
}
\value{
markov_chain
}
\description{
Add an edge to the chain
}
\examples{
m = new_markov_chain(number_of_states = 3)
m = add_edge_by_indices(m, 1, 2, prob=0.2)
print(m)
}
