% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_helpers.R
\name{add_edge_by_labels}
\alias{add_edge_by_labels}
\title{add edge, but from and to are string labels}
\usage{
add_edge_by_labels(m, from, to, ...)
}
\arguments{
\item{m}{markov_chain}

\item{from}{label}

\item{to}{label}

\item{...}{}
}
\value{
markov_chain
}
\description{
add edge, but from and to are string labels
}
\examples{
m = new_markov_chain()
m = add_states(m, c("a", "b", "c"))
m = add_edge_by_labels(m, "b", "c", prob=0.4)
print(m)
}
