% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_class.R
\name{add_state}
\alias{add_state}
\title{Adds a state to a markov_chain}
\usage{
add_state(m, label = NA, x = NA, y = NA, color = NA, shape = NA)
}
\arguments{
\item{m}{markov_chain}

\item{label}{state's label}

\item{x}{x-coordinate (for plotting)}

\item{y}{x-coordinate (for plotting)}

\item{color}{(for plotting)}
}
\value{
markov_chain object
}
\description{
Adds a state to a markov_chain
}
\examples{
m = new_markov_chain()
m = add_state(m,"a", x=1, y=2, color = "reddish")
print(m)
}
