% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_class.R
\name{add_states}
\alias{add_states}
\title{Adds a vector of states to a chain}
\usage{
add_states(m, states, ...)
}
\arguments{
\item{m}{markov_chain}

\item{states}{a vector of state labels}

\item{...}{}
}
\value{

}
\description{
Adds a vector of states to a chain
}
\examples{
m = new_markov_chain()
m = add_states(m, c("a","b","c"), color = "yellow")
print(m)
}
