% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{all_paths}
\alias{all_paths}
\title{Output all paths of a random walk}
\usage{
all_paths(T, nozero = FALSE)
}
\arguments{
\item{T}{time horizon}

\item{nozero}{whether to remove the leading zero}
}
\value{
list
}
\description{
Output all paths of a random walk
}
