% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{attached_cycles}
\alias{attached_cycles}
\title{Markov chain with two attached cycles}
\usage{
attached_cycles(n1 = 4, n2 = 6, pA = 0.4, tail = FALSE)
}
\arguments{
\item{n1}{size of the first cycle}

\item{n2}{size of the second cycle}

\item{pA}{probability of choosing cycle 1}

\item{tail}{whether there is a short tail}
}
\value{
mc
}
\description{
Markov chain with two attached cycles
}
