% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_layout.R
\name{autoscale}
\alias{autoscale}
\title{Scales the layout to mean=0, var=1}
\usage{
autoscale(m, ...)
}
\arguments{
\item{m}{a markov_chain object}

\item{...}{additional arguments to builtin "scale"}
}
\value{
a markov_chain object
}
\description{
Scales the layout to mean=0, var=1
}
