% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_compute.R
\name{classify}
\alias{classify}
\title{Classifies states of the chain}
\usage{
classify(m)
}
\arguments{
\item{m}{a markov_chain object}
}
\value{
a markov_chain object
}
\description{
Classifies states of the chain
}
