% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{color_tester_mc}
\alias{color_tester_mc}
\title{An mc used to test color schemes}
\usage{
color_tester_mc(number_of_states)
}
\arguments{
\item{number_of_states}{}
}
\value{

}
\description{
An mc used to test color schemes
}
\examples{
m = color_tester_mc(10)
m = set_fancy_edge_colors(m)
plot(m)
}
