% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_layout.R
\name{curve_overlapping_edges}
\alias{curve_overlapping_edges}
\title{Make overlapping edges avoid each other}
\usage{
curve_overlapping_edges(m, curve = default$edge.curve)
}
\arguments{
\item{m}{a markov_chain object}

\item{curve}{the curvature}
}
\value{
a markov_chain object
}
\description{
Make overlapping edges avoid each other
}
