% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{deterministically_monotone}
\alias{deterministically_monotone}
\title{Deterministically monotone markov chain}
\usage{
deterministically_monotone(T = 6)
}
\arguments{
\item{T}{the position of the state cut-off ("...")}
}
\value{
a markov_chain object
}
\description{
Deterministically monotone markov chain
}
