% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{facility}
\alias{facility}
\title{Markov chain for the problem with computers breaking down}
\usage{
facility(p = 0.4)
}
\arguments{
\item{p}{the break-down probability}
}
\value{
a markov_chain object
}
\description{
Markov chain for the problem with computers breaking down
}
