% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{first_maximum_time}
\alias{first_maximum_time}
\title{Computes the first maximum time}
\usage{
first_maximum_time(x, T = 0)
}
\arguments{
\item{x}{vector}

\item{T}{numeric}
}
\value{
integer
}
\description{
Computes the first maximum time
}
\examples{
first_maximum_time(x=c(1,3,2,4,5,6,3,4))
}
