% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{gamblers_ruin}
\alias{gamblers_ruin}
\title{Gambler's ruin markov chain}
\usage{
gamblers_ruin(a = 5, p = 0.25)
}
\arguments{
\item{a}{desired winnings}

\item{p}{probability of winning 1 $}
}
\value{

}
\description{
Gambler's ruin markov chain
}
