% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_class.R
\name{index_pair_to_edge}
\alias{index_pair_to_edge}
\title{The row of the edges data frame from "from" and "to"
Returns 0 if there is not such edge}
\usage{
index_pair_to_edge(m, from, to)
}
\arguments{
\item{to}{}
}
\value{

}
\description{
The row of the edges data frame from "from" and "to"
Returns 0 if there is not such edge
}
\examples{
m = random_markov_chain(3)
j = index_pair_to_edge(m,1,2)
m$edges[j,]
}
