% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_helpers.R
\name{index_to_label}
\alias{index_to_label}
\title{Label of a state given index}
\usage{
index_to_label(m, index)
}
\arguments{
\item{m}{markov_chain}

\item{index}{}
}
\value{
string label
}
\description{
Label of a state given index
}
\examples{
m = new_markov_chain()
m = add_states(m,c("a","b","c"))
print(index_to_label(m, 2))
}
