% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_compute.R
\name{is_classified}
\alias{is_classified}
\title{Check whether the states of the chain have already been classified}
\usage{
is_classified(m)
}
\arguments{
\item{m}{}
}
\value{
logical
}
\description{
Check whether the states of the chain have already been classified
}
