% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_class.R
\name{label_pair_to_edge}
\alias{label_pair_to_edge}
\title{The row of the edges data frame from "from" and "to"}
\usage{
label_pair_to_edge(m, from, to)
}
\arguments{
\item{to}{}
}
\value{

}
\description{
The row of the edges data frame from "from" and "to"
}
