% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_class.R
\name{label_to_index}
\alias{label_to_index}
\title{Index of a state given its label}
\usage{
label_to_index(m, label)
}
\arguments{
\item{m}{markov_chain}

\item{label}{}
}
\value{
integer or 0 if no match
}
\description{
Index of a state given its label
}
\examples{
m = new_markov_chain("Empty chain")
m = add_states(m, c("a","b"))
print(label_to_index(m,"b"))
print(label_to_index(m,"c"))
}
