% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{last_maximum_time}
\alias{last_maximum_time}
\title{Computes the last maximum time}
\usage{
last_maximum_time(x, T = 0)
}
\arguments{
\item{x}{vector}

\item{T}{numeric}
}
\value{
integer
}
\description{
Computes the last maximum time
}
\examples{
last_maximum_time(x=c(1,3,2,4,5,6,3,4))
}
