% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_class.R
\name{length.markov_chain}
\alias{length.markov_chain}
\title{Returns the number of states}
\usage{
\method{length}{markov_chain}(x)
}
\arguments{
\item{m}{markov_chain object}
}
\value{
integer
}
\description{
Returns the number of states
}
\examples{
m = new_markov_chain("Chain with 3 states", number_of_states = 3)
print(length(m))
}
