% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_constructors.R
\name{markov_chain}
\alias{markov_chain}
\title{Construct a markov_chain object}
\usage{
markov_chain(x = 0, ...)
}
\arguments{
\item{x}{The object from which the chain is to be constructed. Can be a single integer, a vector, a matrix or a data frame.}
}
\value{

}
\description{
Construct a markov_chain object
}
