% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_summary.R
\name{P_dput}
\alias{P_dput}
\title{Output the transition matrix in a copy-paste form}
\usage{
P_dput(m)
}
\arguments{
\item{m}{a markov_chain object}
}
\value{
a string
}
\description{
Output the transition matrix in a copy-paste form
}
