% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{pattern}
\alias{pattern}
\title{A pattern-waiting markov chain for an arbitrary pattern}
\usage{
pattern(pattern_string, p = 1/2)
}
\arguments{
\item{pattern_string}{a string of "H" and "T"}

\item{p}{probability of "H"}
}
\value{
a markov_chain object
}
\description{
A pattern-waiting markov chain for an arbitrary pattern
}
