% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{pattern_HHH}
\alias{pattern_HHH}
\title{A Markov chain that tracks the pattern HHH
in a series of coin tosses}
\usage{
pattern_HHH()
}
\value{
a markov_chain object
}
\description{
A Markov chain that tracks the pattern HHH
in a series of coin tosses
}
