% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{pattern_HTH}
\alias{pattern_HTH}
\title{A Markov chain that tracks the pattern HTH
in a series of coin tosses}
\usage{
pattern_HTH()
}
\value{
a markov_chain object
}
\description{
A Markov chain that tracks the pattern HTH
in a series of coin tosses
}
