% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{permutation_test}
\alias{permutation_test}
\title{Checks, graphically, if two vectors are independent}
\usage{
permutation_test(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
a plot
}
\description{
Checks, graphically, if two vectors are independent
}
\examples{
permutation_test( c(1,2,3,4,5), c(2,3,4,5,6))
}
