% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_output.R
\name{plot.markov_chain}
\alias{plot.markov_chain}
\title{Plot method for a markov_chain object}
\usage{
\method{plot}{markov_chain}(x, ...)
}
\arguments{
\item{m}{}
}
\value{

}
\description{
Plot method for a markov_chain object
}
