% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_output.R
\name{print.markov_chain}
\alias{print.markov_chain}
\title{The print method for the markov_chain class}
\usage{
\method{print}{markov_chain}(x)
}
\arguments{
\item{m}{a markov_chain object}
}
\description{
The print method for the markov_chain class
}
