% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{random_mc}
\alias{random_mc}
\title{Builds a mc with random transitions}
\usage{
random_mc(n, p = 0.5)
}
\arguments{
\item{n}{number of states}

\item{p}{a parameter governing the total number of edges}
}
\value{

}
\description{
Builds a mc with random transitions
}
\examples{
m = random_mc(3, p = 0.4)
print(transition_matrix(m))
}
