% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{random_walk}
\alias{random_walk}
\title{Random walk}
\usage{
random_walk(T = 3, p = 0.5, reflecting = FALSE, with_dots = FALSE)
}
\arguments{
\item{T}{the random walk goes from -T to T}

\item{p}{the probability of the step to the right}

\item{reflecting}{whether the states T and -T should be reflecting (otherwise absorbing)}

\item{with_dots}{whether the states T and -T should be replaced by dots (to look good in lecture notes)}
}
\value{
a markov chain object
}
\description{
Random walk
}
\examples{
m = random_walk()
plot(m)
}
