% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_compute.R
\name{recurrent_states}
\alias{recurrent_states}
\title{Returns the vector of transient states}
\usage{
recurrent_states(m)
}
\arguments{
\item{m}{a markov_chain object}
}
\value{
vector is indices
}
\description{
Returns the vector of transient states
}
