% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{reflect_after_hitting}
\alias{reflect_after_hitting}
\title{Reflects trajectory once it hits a level}
\usage{
reflect_after_hitting(x, level)
}
\arguments{
\item{x}{vector}

\item{level}{numeric}
}
\value{
vector
}
\description{
Reflects trajectory once it hits a level
}
\examples{
reflect_after_hitting(x=c(1,2,3,4,5,6,4,3), level = 3)
}
