% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{regime_switching}
\alias{regime_switching}
\title{The regime-switching Markov chain}
\usage{
regime_switching(p12 = 0.4, p21 = 0.7)
}
\arguments{
\item{p12}{probability of the transition 1->2}

\item{p21}{probability of the transition 2->1}
}
\value{
a markov_chain object
}
\description{
The regime-switching Markov chain
}
