% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_layout.R
\name{rotate}
\alias{rotate}
\title{Rotate the layout of a markov_chain}
\usage{
rotate(m, alpha, pca = FALSE)
}
\arguments{
\item{m}{a markov_chain object}

\item{alpha}{angle}

\item{pca}{whether to set angle using principal components}
}
\value{
a markov_chain object
}
\description{
Rotate the layout of a markov_chain
}
