% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_output.R
\name{set_absorbing_state_color}
\alias{set_absorbing_state_color}
\title{Colors all absorbing states}
\usage{
set_absorbing_state_color(m, color = default$absorbing.state.color)
}
\arguments{
\item{m}{a markov_chain object}

\item{color}{a color}
}
\value{
a markov_chain object
}
\description{
Colors all absorbing states
}
