% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_layout.R
\name{set_auto_layout}
\alias{set_auto_layout}
\title{Sets an automatic layout using the specified algorithm}
\usage{
set_auto_layout(m, algorithm = default$layout.algorithm)
}
\arguments{
\item{algorithm}{an igraph layout algorithm}
}
\value{

}
\description{
Sets an automatic layout using the specified algorithm
}
