% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_output.R
\name{set_auto_state_colors}
\alias{set_auto_state_colors}
\title{Colors states according to their classes}
\usage{
set_auto_state_colors(m, by_recurrence = T, by_class = F, ...)
}
\arguments{
\item{m}{a markov_chain object}

\item{by_recurrence}{logical}

\item{by_class}{logical}

\item{...}{additional arguments (palette)}
}
\description{
Colors states according to their classes
}
