% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_helpers.R
\name{set_edge}
\alias{set_edge}
\title{Sets edge properties}
\usage{
set_edge(m, from, to, ...)
}
\arguments{
\item{from}{if integer index, otherwise state label}

\item{to}{if integer index, otherwise state label}

\item{...}{a list of named properties to be set}
}
\value{
markov_chain
}
\description{
Sets edge properties
}
\examples{
m = markov_chain(2)
m = add_edge(m,1,2,prob=0.2)
m = set_edge_(m,1,2, prob=0.3)
print(m$edges)
}
