% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_helpers.R
\name{set_edge_properties_by_indices}
\alias{set_edge_properties_by_indices}
\title{Sets a property of an edge by its indices}
\usage{
set_edge_properties_by_indices(m, from, to, ...)
}
\arguments{
\item{from}{index}

\item{to}{index}

\item{...}{named properties to be set}
}
\value{
markov_chain
}
\description{
Sets a property of an edge by its indices
}
\examples{
m = new_markov_chain("Chain",2)
m = add_edge_by_indices(m,1,2,prob=0.2)
m = set_edge_properties_by_indices(m,1,2, prob=0.3)
print(m$edges)
}
