% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_class.R
\name{set_edge_properties_by_labels}
\alias{set_edge_properties_by_labels}
\title{Sets a property of an edge by its labels}
\usage{
set_edge_properties_by_labels(m, from, to, ...)
}
\arguments{
\item{from}{label}

\item{to}{label}

\item{...}{named properties to be set}
}
\value{
markov_chain
}
\description{
Sets a property of an edge by its labels
}
\examples{
m = new_markov_chain("Chain")
m = add_states(m, c("b", "a"))
m = add_edge_by_labels(m, "a", "b", prob=0.2)
m = set_edge_properties_by_indices(m, "a", "b", prob=0.3)
print(m$edges)
}
