% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_output.R
\name{set_fancy_layout}
\alias{set_fancy_layout}
\title{Sets an automatic layout using the specified algorithm}
\usage{
set_fancy_layout(m, algorithm = pkg.env$default.layout.algorithm)
}
\arguments{
\item{algorithm}{}
}
\value{

}
\description{
Sets an automatic layout using the specified algorithm
}
