% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_output.R
\name{set_graphics_parameters}
\alias{set_graphics_parameters}
\title{Sets graphics parameters of a markov_chain}
\usage{
set_graphics_parameters(m, ...)
}
\arguments{
\item{m}{a markov_chain object}

\item{...}{the parameters to be set}
}
\description{
Sets graphics parameters of a markov_chain
}
