% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_helpers.R
\name{set_state}
\alias{set_state}
\title{Sets properties of a state}
\usage{
set_state(m, x, ...)
}
\arguments{
\item{m}{a markov_chain object}

\item{x}{if integer, state index, otherwise state label}

\item{...}{named properties to be set}
}
\value{
markov_chain
}
\description{
Sets properties of a state
}
\examples{
m = markov_chain(3)
m = set_state(m,c(1,3), color="Aquamarine")
print(m$states)
}
