% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_class.R
\name{set_state_properties_by_index}
\alias{set_state_properties_by_index}
\title{Sets a property of a state by its index}
\usage{
set_state_properties_by_index(m, index, ...)
}
\arguments{
\item{m}{markov_chain}

\item{index}{integer}

\item{...}{named properties to be set}
}
\value{
markov_chain
}
\description{
Sets a property of a state by its index
}
\examples{
m = new_markov_chain("Chain",2)
m = set_state_properties_by_index(m,1, color="Aquamarine")
print(m$states)
}
