% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_helpers.R
\name{set_state_properties_by_label}
\alias{set_state_properties_by_label}
\title{Sets a property of a state by its label}
\usage{
set_state_properties_by_label(m, label, ...)
}
\arguments{
\item{m}{markov_chain}

\item{label}{string}

\item{...}{named properties to be set}
}
\value{
markov_chain
}
\description{
Sets a property of a state by its label
}
\examples{
m = new_markov_chain("Chain")
m = add_states(m,c("a","b"))
m = set_state_properties_by_index(m,"a", color="Aquamarine")
print(m$states)
}
