% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_layout.R
\name{shift}
\alias{shift}
\title{Shift the layout of a markov_chain}
\usage{
shift(m, ax, ay = 0)
}
\arguments{
\item{m}{a markov_chain object}

\item{ax}{shift amount in x}

\item{ay}{shift amount in y}
}
\value{
a markov_chain object
}
\description{
Shift the layout of a markov_chain
}
