% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{simulate_walk}
\alias{simulate_walk}
\title{Simulate a data frame of paths of a random walk}
\usage{
simulate_walk(nsim, T, p = 0.5)
}
\arguments{
\item{nsim}{number of trajectories simulated}

\item{T}{time horizon}

\item{p}{up probability (defaults to 0.5)}
}
\value{
data.framelibrary
}
\description{
Simulate a data frame of paths of a random walk
}
\examples{
simulate_walk(nsim=5, T=10, p=0.3)
}
