% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{single_bridge}
\alias{single_bridge}
\title{A single trajectory of a random bridge}
\usage{
single_bridge(T, k, nozero = FALSE)
}
\arguments{
\item{T}{time horizon}

\item{k}{value at T}

\item{nozero}{whether to remove the leading zero}
}
\value{
vector
}
\description{
Does not return the initial 0
}
\examples{
single_bridge(T=15, k=3, nozero=TRUE)
}
