% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{single_trajectory}
\alias{single_trajectory}
\title{Simulate a single trajectory of a random walk}
\usage{
single_trajectory(T, p = 0.5, nozero = FALSE)
}
\arguments{
\item{T}{time horizon}

\item{p}{up probability (defaults to 0.5)}

\item{nozero}{whether to remove the leading zero}
}
\value{
vector
}
\description{
Simulate a single trajectory of a random walk
}
\examples{
single_trajectory(T=10, p=0.3, nozero=FALSE)
}
