% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_layout.R
\name{stretch}
\alias{stretch}
\title{Stretch (scale) the layout of a markov_chain}
\usage{
stretch(m, ax, ay = ax)
}
\arguments{
\item{m}{a markov_chain object}

\item{ax}{stretch factor in x}

\item{ay}{stretch factor in y}
}
\value{
a markov_chain object
}
\description{
Stretch (scale) the layout of a markov_chain
}
