% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_summary.R
\name{summary.markov_chain}
\alias{summary.markov_chain}
\title{The summary method for the markov_chain class}
\usage{
\method{summary}{markov_chain}(object, ...)
}
\arguments{
\item{m}{a markov_chain object}
}
\description{
The summary method for the markov_chain class
}
