% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{tennis}
\alias{tennis}
\title{The tennis chain}
\usage{
tennis(p = 0.4)
}
\arguments{
\item{p}{probability of winning for player 1 in a single rally}
}
\value{
a markov_chain object
}
\description{
The tennis chain
}
