% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_compute.R
\name{transition_matrix}
\alias{transition_matrix}
\title{Returns the transition matrix}
\usage{
transition_matrix(m)
}
\value{

}
\description{
Returns the transition matrix
}
\examples{
df = data_frame(
  from = c(1,2,2,3),
  to = c(2,1,3,1),
  prob = c(1, 0.3, 0.7, 1)
)
m = new_mc_from_data_frame(df)
print(transition_matrix(m))
}
