% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_helpers.R
\name{add_state}
\alias{add_state}
\title{Adds a state to a markov_chain}
\usage{
add_state(m, l = NA, ...)
}
\arguments{
\item{m}{a markov_chain object}

\item{...}{Extra arguments like color, x, y, etc.}

\item{v}{a vector. States with labels in v (cast to strings) are added. If length(v)==1
and v is numeric, v states with labels "1", ..., "v" are added.}
}
\value{
markov_chain object
}
\description{
Adds a state to a markov_chain
}
\examples{
m = new_markov_chain(0)
m = add_state(m,c("a","b","c"), x=1, y=2, color = "reddish")
print(m)
}
