% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_helpers.R
\name{add_state}
\alias{add_state}
\title{Adds a state to a markov_chain}
\usage{
add_state(m, l = NA, ...)
}
\arguments{
\item{m}{a markov_chain object}

\item{l}{a vector. States with labels in l (cast to strings) are added. If length(l)==1
and l is numeric, l states with labels "1", ..., "l" are added.}

\item{...}{Extra arguments like color, x, y, etc.}
}
\value{
markov_chain object
}
\description{
Adds a state to a markov_chain
}
