% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_class.R
\name{curve_overlapping_edges}
\alias{curve_overlapping_edges}
\title{Make overlapping edges avoid each other}
\usage{
curve_overlapping_edges(m, curve = pkg.env$default.curve)
}
\arguments{
\item{curve}{the curvature}
}
\value{

}
\description{
Make overlapping edges avoid each other
}
\examples{
P = matrix( rep(1/3,9), ncol=3)
m = new_markov_chain_from_matrix(P)
}
