% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{first_hitting_time}
\alias{first_hitting_time}
\title{Computes the first hitting time (>=)}
\usage{
first_hitting_time(x, level, T = 0)
}
\arguments{
\item{x}{vector}

\item{level}{numeric}

\item{T}{numeric}
}
\value{
integer
}
\description{
Computes the first hitting time (>=)
}
