% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_helpers.R
\name{label_to_index}
\alias{label_to_index}
\title{Index of a state given its label}
\usage{
label_to_index(m, label)
}
\arguments{
\item{m}{a markov_chain object}

\item{label}{string.}
}
\value{
integer or 0 if label does not match
}
\description{
Index of a state given its label
}
\examples{
m = new_markov_chain(c("a","b"))
print(label_to_index(m,"b"))
print(label_to_index(m,"c"))
}
