% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_class.R
\name{nedges}
\alias{nedges}
\title{Return the number of edges}
\usage{
nedges(m)
}
\arguments{
\item{m}{markov_chain object}
}
\value{
integer
}
\description{
Return the number of edges
}
\examples{
m = new_markov_chain("Chain with 3 states", number_of_states = 3)
m = add_edge_by_labels("1","2")
m = add_edge_by_labels("1","3")
print(nedges(m))
}
