% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_helpers.R
\name{nedges}
\alias{nedges}
\title{Return the number of edges}
\usage{
nedges(m)
}
\arguments{
\item{m}{a markov_chain object}
}
\value{
integer
}
\description{
Return the number of edges
}
\examples{
m = markov_chain(3) \%>\%
  add_edge(1,2) \%>\%
  add_edge(1,3)
print(nedges(m))
}
