% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_output.R
\name{plot.markov_chain}
\alias{plot.markov_chain}
\title{Plot method for a markov_chain object}
\usage{
\method{plot}{markov_chain}(x, ...)
}
\arguments{
\item{x}{the markov_chain object to be plotted}

\item{...}{additional graphics parameters}
}
\description{
Plot method for a markov_chain object
}
