% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_examples.R
\name{professor}
\alias{professor}
\title{A Markov Chain for the problem with the professor and umbrellas}
\usage{
professor(p_morning = 0.05, p_afternoon = 0.2)
}
\arguments{
\item{p_morning}{probability of getting wet in the morning}

\item{p_afternoon}{probability of getting wet in the afternoon}
}
\value{
a markov_chain object
}
\description{
A Markov Chain for the problem with the professor and umbrellas
}
