% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_output.R
\name{set_auto_edge_colors}
\alias{set_auto_edge_colors}
\title{Colors edges according to their probabilities}
\usage{
set_auto_edge_colors(m, discrete = T, nbins = 3, ...)
}
\arguments{
\item{discrete}{logical, distinguishable vs. in order}

\item{nbins}{if discrete, how many categories}

\item{...}{}
}
\value{

}
\description{
Colors edges according to their probabilities
}
\examples{
m = color_tester_markov_chain(15)
m = set_fancy_edge_colors(m)
plot(m)
}
