% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_output.R
\name{set_graphics_parameters}
\alias{set_graphics_parameters}
\title{Sets graphics parameters of a markov_chain}
\usage{
set_graphics_parameters(m, ...)
}
\arguments{
\item{...}{the parameters to be set}
}
\value{

}
\description{
Sets graphics parameters of a markov_chain
}
\examples{
m = random_markov_chain(3)
plot(m)
m = set_graphics_parameters(edge.label = NA)
plot(m)
}
