% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_compute.R
\name{transition_matrix}
\alias{transition_matrix}
\title{Returns the transition matrix}
\usage{
transition_matrix(m)
}
\arguments{
\item{m}{a markov_chain object}
}
\value{
a matrix
}
\description{
Returns the transition matrix
}
