% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{univariate_table}
\alias{univariate_table}
\title{Create a univariate distribution table}
\usage{
univariate_table(x, p)
}
\arguments{
\item{x}{vector of possible values}

\item{p}{vector of their probabilities}
}
\value{
html?
}
\description{
Create a table using kableExtra
}
\examples{
univariate_table(c(1,2,3), c(0.1, 0.3, 0.6))
}
